pyproject.toml
[project]
name = "cv-match"
version = "0.1.0"
description = "Local CLI experiment for deterministic multi-round resume matching with Pydantic AI"
readme = "README.md"
authors = [
    { name = "frankqdwang", email = "wang15850317595@163.com" }
]
requires-python = ">=3.12"
dependencies = [
    "httpx>=0.28.1",
    "pydantic-ai>=1.73.0",
    "pydantic-settings>=2.13.1",
    "python-dotenv>=1.2.2",
    "pyyaml>=6.0.3",
]

[dependency-groups]
dev = [
    "pytest>=8.4.2",
]

[project.scripts]
cv-match = "cv_match.cli:main"

[build-system]
requires = ["uv_build>=0.11.2,<0.12.0"]
build-backend = "uv_build"

[tool.pytest.ini_options]
pythonpath = ["src"]
